src/requirement.txt
boto==2.49.0
boto3==1.14.3
botocore==1.17.3
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
Django==2.2
django-environ==0.4.5
django-rest-knox==4.1.0
djangorestframework==3.11.0
docutils==0.15.2
dynaconf==2.2.3
environ==1.0
hvac==0.10.3
idna==2.9
Jinja2==2.11.2
jmespath==0.10.0
knox==0.1.0
loguru==0.5.1
MarkupSafe==1.1.1
psycopg2==2.8.5
pycparser==2.20
python-box==3.4.6
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
win32-setctime==1.0.1
